refactor(api): promote WriteOutcome/outcome_response/resolve_actor to the shared writes module (#419)
So domains B/C/D import the write-result helpers from one place instead of each
re-declaring them. No behavior change; pure move + visibility (WriteOutcome is
pub for the external harness).
